Extract adding of organizations/domains functionality to own file
