+++ scripts/domain-loader/src/align-organizations-domains.js
@@ -1,0 +1,15 @@
+const {
+  checkClaimCount,
+  checkDomain,
+  checkOrganization,
+  createClaim,
+  createDomain,
+  createOrganization,
+  slugify,
+} = require('./src')
+
+const alignOrganizationsDomains = () => {}
+
+module.exports = {
+  alignOrganizationsDomains,
+}
