+++ scripts/domain-loader/src/helpers/check-claim-count.js
