+++ scripts/domain-loader/src/helpers/check-domain.js
