+++ scripts/domain-loader/src/helpers/create-claim.js
