+++ scripts/domain-loader/src/helpers/create-domain.js
