+++ scripts/domain-loader/src/helpers/slugify.js
