+++ scripts/domain-loader/src/index.js
@@ -1,17 +1,7 @@
-const { checkClaimCount } = require('./check-claim-count')
-const { checkDomain } = require('./check-domain')
-const { checkOrganization } = require('./check-organization')
-const { createClaim } = require('./create-claim')
-const { createDomain } = require('./create-domain')
-const { createOrganization } = require('./create-organization')
-const slugify = require('./slugify')
+const { addOrganizationsDomains } = require('../add-organizations-domains')
+const { alignOrganizationsDomains } = require('../align-organizations-domains')
 
 module.exports = {
-  checkClaimCount,
-  checkDomain,
-  checkOrganization,
-  createClaim,
-  createDomain,
-  createOrganization,
-  ...slugify,
+  addOrganizationsDomains,
+  alignOrganizationsDomains,
 }
