--- scripts/domain-loader/src/setupEnv.js
@@ -1,4 +1,0 @@
-require('dotenv-safe').config({
-  example: './.env.example',
-  path: './test.env',
-})
